Big ol commit consisting of most of the work from last week.
Fixes up some things that were broken today. In particular, an
off-by-one bug in the initial inserting piece was fixed and a bug with
 querying a single cell multiple times was fixed. DynamoDB support is
added though it's a little shakey since the DynamoDB data model is a bit
stranger than hbase's and atomicity of updates are still uncertain.

Project is cleaned up significantly, repl-style tests are confined to
the tests.clj file.